+++ src/ast/expression.rs
@@ -1,66 +1,5 @@
+use super::{BlockStatement, Token};
 use std::fmt::Display;
-
-#[derive(Debug, PartialEq, Clone)]
-pub struct TokenSpan {
-    pub line: usize,
-    pub column: usize,
-    pub length: usize,
-}
-
-#[derive(Debug, PartialEq, Clone)]
-pub enum TokenKind {
-    // special
-    WHITESPACE,
-    ILLEGAL,
-    EOF,
-
-    // operators
-    ASSIGN,
-    PLUS,
-    MINUS,
-    ASTERISK,
-    SLASH,
-    BANG,
-    EQ,
-    NEQ,
-    LT,
-    GT,
-    LTE,
-    GTE,
-
-    // delimiters
-    COMMA,
-    SEMICOLON,
-
-    // brackets
-    LPAREN,
-    RPAREN,
-    LBRACE,
-    RBRACE,
-    RBRACKET,
-    LBRACKET,
-
-    // literals
-    INT,
-    STRING,
-
-    // keywords & identifiers
-    LET,
-    FUNCTION,
-    RETURN,
-    IF,
-    ELSE,
-    TRUE,
-    FALSE,
-    IDENT,
-}
-
-#[derive(Debug, PartialEq, Clone)]
-pub struct Token {
-    pub span: TokenSpan,
-    pub kind: TokenKind,
-    pub literal: Box<str>,
-}
 
 #[derive(Debug, PartialEq, Clone)]
 pub struct IdentExpression {
@@ -283,105 +222,6 @@ impl Display for Expression {
             Expression::CALL(expression) => write!(f, "{}", expression),
             Expression::PREFIX(expression) => write!(f, "{}", expression),
             Expression::INFIX(expression) => write!(f, "{}", expression),
-        }
-    }
-}
-
-#[derive(Debug, PartialEq, Clone)]
-pub struct BlockStatement {
-    pub token: Token,
-    pub statements: Vec<Statement>,
-}
-
-impl Display for BlockStatement {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let mut output = String::new();
-
-        for statement in &self.statements {
-            output.push_str(&format!("{}", statement));
-        }
-
-        write!(f, "{{{}}}", output)
-    }
-}
-
-#[derive(Debug, PartialEq, Clone)]
-pub struct LetStatement {
-    pub token: Token,
-    pub identifier: Box<str>,
-    pub expression: Expression,
-}
-
-impl Display for LetStatement {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "let {} = {};", self.identifier, self.expression)
-    }
-}
-
-#[derive(Debug, PartialEq, Clone)]
-pub struct ReturnStatement {
-    pub token: Token,
-    pub expression: Expression,
-}
-
-impl Display for ReturnStatement {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "return {};", self.expression)
-    }
-}
-
-#[derive(Debug, PartialEq, Clone)]
-pub struct ExpressionStatement {
-    pub token: Token,
-    pub expression: Expression,
-}
-
-impl Display for ExpressionStatement {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{};", self.expression)
-    }
-}
-
-#[derive(Debug, PartialEq, Clone)]
-pub enum Statement {
-    Block(BlockStatement),
-    Let(LetStatement),
-    Return(ReturnStatement),
-    Expression(ExpressionStatement),
-}
-
-impl Display for Statement {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Statement::Block(statement) => write!(f, "{}", statement),
-            Statement::Let(statement) => write!(f, "{}", statement),
-            Statement::Return(statement) => write!(f, "{}", statement),
-            Statement::Expression(statement) => write!(f, "{}", statement),
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct Program {
-    pub statements: Vec<Statement>,
-}
-
-impl Display for Program {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let mut output = String::new();
-
-        for statement in &self.statements {
-            output.push_str(&format!("{}", statement));
-        }
-
-        write!(f, "{}", output)
-    }
-}
-
-impl Program {
-    pub fn new() -> Self {
-        Program {
-            statements: Vec::new(),
         }
     }
 }
