+++ src/ast/statement.rs
@@ -1,0 +1,76 @@
+use super::{Expression, Token};
+use std::fmt::Display;
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct BlockStatement {
+    pub token: Token,
+    pub statements: Vec<Statement>,
+}
+
+impl Display for BlockStatement {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let mut output = String::new();
+
+        for statement in &self.statements {
+            output.push_str(&format!("{}", statement));
+        }
+
+        write!(f, "{{{}}}", output)
+    }
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct LetStatement {
+    pub token: Token,
+    pub identifier: Box<str>,
+    pub expression: Expression,
+}
+
+impl Display for LetStatement {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "let {} = {};", self.identifier, self.expression)
+    }
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct ReturnStatement {
+    pub token: Token,
+    pub expression: Expression,
+}
+
+impl Display for ReturnStatement {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "return {};", self.expression)
+    }
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct ExpressionStatement {
+    pub token: Token,
+    pub expression: Expression,
+}
+
+impl Display for ExpressionStatement {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{};", self.expression)
+    }
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub enum Statement {
+    Block(BlockStatement),
+    Let(LetStatement),
+    Return(ReturnStatement),
+    Expression(ExpressionStatement),
+}
+
+impl Display for Statement {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Statement::Block(statement) => write!(f, "{}", statement),
+            Statement::Let(statement) => write!(f, "{}", statement),
+            Statement::Return(statement) => write!(f, "{}", statement),
+            Statement::Expression(statement) => write!(f, "{}", statement),
+        }
+    }
+}
